editoast: extract valkey utils in new crate `cache`
