+++ editoast/Cargo.toml
@@ -8,6 +8,7 @@ edition.workspace = true
 members = [
   ".",
   "authz",
+  "cache",
   "common",
   "core_client",
   "database",
@@ -32,8 +33,10 @@ license = "LGPL-3.0"
 
 [workspace.dependencies]
 approx = "0.5.1"
+arcstr = { version = "1.2.0", features = ["serde"] }
 async-std = { version = "1.13.2", features = ["attributes", "tokio1"] }
 authz = { path = "./authz" }
+cache = { path = "./cache" }
 chrono = { version = "0.4.42", default-features = false, features = ["serde"] }
 clap = { version = "4.5", features = ["derive", "env"] }
 common = { path = "./common" }
@@ -44,6 +47,10 @@ deadpool = { version = "0.12.3", features = [
   "managed",
   "rt_tokio_1",
   "unmanaged",
+] }
+deadpool-redis = { version = "0.22.0", features = [
+  "tokio-comp",
+  "tokio-native-tls-comp",
 ] }
 derive_more = { version = "2.0.1", features = [
   "debug",
@@ -130,7 +137,7 @@ uuid = { version = "1.18.1", features = ["serde", "v4"] }
 [dependencies]
 anyhow = "1.0"
 approx.workspace = true
-arcstr = { version = "1.2.0", features = ["serde"] }
+arcstr.workspace = true
 authz.workspace = true
 axum = { version = "0.8.4", default-features = false, features = [
   "multipart",
@@ -143,6 +150,7 @@ axum-test = { version = "18.0", default-features = false }
 axum-tracing-opentelemetry = { version = "0.30.0", default-features = false, features = [
   "tracing_level_info",
 ] }
+cache.workspace = true
 chrono.workspace = true
 clap.workspace = true
 colored = "3.0.0"
@@ -151,10 +159,7 @@ core_client = { workspace = true, features = ["mocking_client"] }
 dashmap = "6.1.0"
 database.workspace = true
 deadpool.workspace = true
-deadpool-redis = { version = "0.22.0", features = [
-  "tokio-comp",
-  "tokio-native-tls-comp",
-] }
+deadpool-redis.workspace = true
 derive_more.workspace = true
 diesel.workspace = true
 diesel-async = { workspace = true }
@@ -185,9 +190,6 @@ json-patch = { version = "4.0.0", default-features = false, features = [
 ] }
 lapin.workspace = true
 linkme.workspace = true
-lz4_flex = { version = "0.11.5", default-features = false, features = [
-  "frame",
-] }
 mime = "0.3.17"
 mvt.workspace = true
 opentelemetry.workspace = true
