+++ editoast/cache/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "cache"
+license.workspace = true
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+arcstr.workspace = true
+deadpool.workspace = true
+deadpool-redis.workspace = true
+futures.workspace = true
+lz4_flex = { version = "0.11.5", default-features = false, features = [
+  "frame",
+] }
+rand.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tracing.workspace = true
+url.workspace = true
