+++ editoast/src/main.rs
@@ -7,7 +7,6 @@ mod generated_data;
 mod infra_cache;
 mod map;
 mod models;
-mod valkey_utils;
 mod views;
 
 // HACK: temporary while moving models in editoast_models to be able to resolve the `crate::prelude` generated by the Model macro
@@ -40,13 +39,13 @@ use database::DbConnectionPoolV2;
 use tracing_subscriber::util::SubscriberInitExt;
 pub use views::AppState;
 
+pub use cache::ValkeyClient;
+pub use cache::ValkeyConnection;
 use opentelemetry_otlp::WithExportConfig as _;
 use std::io::IsTerminal;
 use std::process::exit;
 use std::sync::Arc;
 use tracing::error;
-pub use valkey_utils::ValkeyClient;
-pub use valkey_utils::ValkeyConnection;
 
 /// The mode editoast is running in
 ///
