+++ editoast/src/views/path/pathfinding.rs
@@ -36,11 +36,11 @@ use crate::error::InternalError;
 use crate::error::Result;
 use crate::models::Infra;
 use crate::models::RollingStock;
-use crate::valkey_utils::ValkeyConnection;
 use crate::views::AuthenticationExt;
 use crate::views::AuthorizationError;
 use crate::views::path::PathfindingError;
 use crate::views::path::path_item_cache::PathItemCache;
+use cache::ValkeyConnection;
 use editoast_models::prelude::*;
 
 /// Path input is described by some rolling stock information
