+++ editoast/src/views/test_app.rs
@@ -41,8 +41,8 @@ use crate::generated_data::speed_limit_tags_config::SpeedLimitTagIds;
 use crate::infra_cache::InfraCache;
 use crate::map::MapLayers;
 use crate::models::PgAuthDriver;
-use crate::valkey_utils::ValkeyConfig;
 use crate::views::service_router;
+use cache::ValkeyConfig;
 use fga::client::DEFAULT_OPENFGA_MAX_CHECKS_PER_BATCH_CHECK;
 use fga::client::DEFAULT_OPENFGA_MAX_TUPLES_PER_WRITE;
